All tests updated
