+++ tests/fixtures/stats.py
@@ -1,19 +1,9 @@
-"""Define tests for stat endpoints."""
-# pylint: disable=redefined-outer-name,too-many-arguments
-
-import datetime
-import json
-
-import aiohttp
+"""Define fixtures related to the "dailystats" endpoint."""
 import pytest
 
-from regenmaschine import Client
 
-from .const import TEST_HOST, TEST_PORT
-
-
-@pytest.fixture(scope='module')
-def fixture_on_date():
+@pytest.fixture()
+def dailystats_date_json():
     """Return a /dailystats/<DATE> response."""
     return {
         "id": 0,
@@ -29,24 +19,18 @@ def fixture_on_date():
     }
 
 
-@pytest.fixture(scope='module')
-def fixture_upcoming():
+@pytest.fixture()
+def dailystats_details_json():
     """Return a /dailystats/details response."""
     return {
         "DailyStatsDetails": [{
-            "dayTimestamp":
-                1528092000,
-            "day":
-                "2018-06-04",
-            "mint":
-                12.779999999999999,
-            "maxt":
-                33.329999999999998,
-            "icon":
-                2,
+            "dayTimestamp": 1528092000,
+            "day": "2018-06-04",
+            "mint": 12.779999999999999,
+            "maxt": 33.329999999999998,
+            "icon": 2,
             "programs": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -65,8 +49,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    2,
+                "id": 2,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -85,8 +68,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    3,
+                "id": 3,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -97,8 +79,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    4,
+                "id": 4,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -110,8 +91,7 @@ def fixture_upcoming():
                 }]
             }],
             "simulatedPrograms": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 2,
                     "scheduledWateringTime": 300,
@@ -122,19 +102,13 @@ def fixture_upcoming():
                 }]
             }]
         }, {
-            "dayTimestamp":
-                1528178400,
-            "day":
-                "2018-06-05",
-            "mint":
-                13.890000000000001,
-            "maxt":
-                34.439999999999998,
-            "icon":
-                3,
+            "dayTimestamp": 1528178400,
+            "day": "2018-06-05",
+            "mint": 13.890000000000001,
+            "maxt": 34.439999999999998,
+            "icon": 3,
             "programs": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -153,8 +127,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    2,
+                "id": 2,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -173,8 +146,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    3,
+                "id": 3,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -185,8 +157,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    4,
+                "id": 4,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -198,8 +169,7 @@ def fixture_upcoming():
                 }]
             }],
             "simulatedPrograms": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 2,
                     "scheduledWateringTime": 300,
@@ -210,19 +180,13 @@ def fixture_upcoming():
                 }]
             }]
         }, {
-            "dayTimestamp":
-                1528264800,
-            "day":
-                "2018-06-06",
-            "mint":
-                14.44,
-            "maxt":
-                32.780000000000001,
-            "icon":
-                3,
+            "dayTimestamp": 1528264800,
+            "day": "2018-06-06",
+            "mint": 14.44,
+            "maxt": 32.780000000000001,
+            "icon": 3,
             "programs": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -241,8 +205,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    2,
+                "id": 2,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -262,8 +225,7 @@ def fixture_upcoming():
                 }]
             }],
             "simulatedPrograms": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 2,
                     "scheduledWateringTime": 300,
@@ -274,19 +236,13 @@ def fixture_upcoming():
                 }]
             }]
         }, {
-            "dayTimestamp":
-                1528351200,
-            "day":
-                "2018-06-07",
-            "mint":
-                13.890000000000001,
-            "maxt":
-                33.890000000000001,
-            "icon":
-                3,
+            "dayTimestamp": 1528351200,
+            "day": "2018-06-07",
+            "mint": 13.890000000000001,
+            "maxt": 33.890000000000001,
+            "icon": 3,
             "programs": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -305,8 +261,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    2,
+                "id": 2,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -326,8 +281,7 @@ def fixture_upcoming():
                 }]
             }],
             "simulatedPrograms": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 2,
                     "scheduledWateringTime": 300,
@@ -338,19 +292,13 @@ def fixture_upcoming():
                 }]
             }]
         }, {
-            "dayTimestamp":
-                1528437600,
-            "day":
-                "2018-06-08",
-            "mint":
-                13.890000000000001,
-            "maxt":
-                34.439999999999998,
-            "icon":
-                3,
+            "dayTimestamp": 1528437600,
+            "day": "2018-06-08",
+            "mint": 13.890000000000001,
+            "maxt": 34.439999999999998,
+            "icon": 3,
             "programs": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -369,8 +317,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    2,
+                "id": 2,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -390,8 +337,7 @@ def fixture_upcoming():
                 }]
             }],
             "simulatedPrograms": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 2,
                     "scheduledWateringTime": 300,
@@ -402,19 +348,13 @@ def fixture_upcoming():
                 }]
             }]
         }, {
-            "dayTimestamp":
-                1528524000,
-            "day":
-                "2018-06-09",
-            "mint":
-                14.44,
-            "maxt":
-                33.890000000000001,
-            "icon":
-                3,
+            "dayTimestamp": 1528524000,
+            "day": "2018-06-09",
+            "mint": 14.44,
+            "maxt": 33.890000000000001,
+            "icon": 3,
             "programs": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -433,8 +373,7 @@ def fixture_upcoming():
                     "wateringFlag": 0
                 }]
             }, {
-                "id":
-                    2,
+                "id": 2,
                 "zones": [{
                     "id": 1,
                     "scheduledWateringTime": 1243,
@@ -454,8 +393,7 @@ def fixture_upcoming():
                 }]
             }],
             "simulatedPrograms": [{
-                "id":
-                    1,
+                "id": 1,
                 "zones": [{
                     "id": 2,
                     "scheduledWateringTime": 300,
@@ -475,32 +413,3 @@ def fixture_upcoming():
             "simulatedPrograms": []
         }]
     }
-
-
-@pytest.mark.asyncio
-async def test_endpoints(
-        aresponses, fixture_on_date, fixture_upcoming, event_loop):
-    """Test all endpoints."""
-    today = datetime.date.today()
-    today_str = today.strftime('%Y-%m-%d')
-
-    aresponses.add(
-        '{0}:{1}'.format(TEST_HOST, TEST_PORT),
-        '/api/4/dailystats/{0}'.format(today_str), 'get',
-        aresponses.Response(text=json.dumps(fixture_on_date), status=200))
-    aresponses.add(
-        '{0}:{1}'.format(TEST_HOST, TEST_PORT), '/api/4/dailystats/details',
-        'get',
-        aresponses.Response(text=json.dumps(fixture_upcoming), status=200))
-
-    # pylint: disable=protected-access
-    async with aiohttp.ClientSession(loop=event_loop) as websession:
-        client = Client(TEST_HOST, websession, port=TEST_PORT, ssl=False)
-        client._authenticated = True
-        client._access_token = '12345'
-
-        data = await client.stats.on_date(today)
-        assert data['percentage'] == 100
-
-        data = await client.stats.upcoming(details=True)
-        assert len(data[0]['programs']) == 4
