+++ tests/fixtures/watering.py
@@ -1,35 +1,20 @@
-"""Define tests for restriction endpoints."""
-# pylint: disable=redefined-outer-name,too-many-arguments
-
-import datetime
-import json
-
-import aiohttp
+"""Define fixtures related to the "watering" endpoint."""
 import pytest
 
-from regenmaschine import Client
 
-from .const import TEST_HOST, TEST_PORT
-
-
-@pytest.fixture(scope='module')
-def fixture_log():
+@pytest.fixture()
+def watering_log_json():
     """Return a /watering/log/details/<DATE>/<DAYS> response."""
     return {
         "waterLog": {
             "days": [{
-                "date":
-                    "2018-06-01",
-                "dateTimestamp":
-                    1527832800,
+                "date": "2018-06-01",
+                "dateTimestamp": 1527832800,
                 "programs": [{
-                    "id":
-                        1,
+                    "id": 1,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            0,
+                        "uid": 1,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-01 06:00:02",
@@ -39,10 +24,8 @@ def fixture_log():
                             "realDuration": 1243
                         }]
                     }, {
-                        "uid":
-                            2,
-                        "flag":
-                            0,
+                        "uid": 2,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-01 06:20:45",
@@ -53,13 +36,10 @@ def fixture_log():
                         }]
                     }]
                 }, {
-                    "id":
-                        2,
+                    "id": 2,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            0,
+                        "uid": 1,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-01 22:00:01",
@@ -69,10 +49,8 @@ def fixture_log():
                             "realDuration": 1243
                         }]
                     }, {
-                        "uid":
-                            2,
-                        "flag":
-                            0,
+                        "uid": 2,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-01 22:20:44",
@@ -84,18 +62,13 @@ def fixture_log():
                     }]
                 }]
             }, {
-                "date":
-                    "2018-06-02",
-                "dateTimestamp":
-                    1527919200,
+                "date": "2018-06-02",
+                "dateTimestamp": 1527919200,
                 "programs": [{
-                    "id":
-                        1,
+                    "id": 1,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            0,
+                        "uid": 1,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 06:00:01",
@@ -105,10 +78,8 @@ def fixture_log():
                             "realDuration": 1217
                         }]
                     }, {
-                        "uid":
-                            2,
-                        "flag":
-                            0,
+                        "uid": 2,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 06:20:18",
@@ -119,13 +90,10 @@ def fixture_log():
                         }]
                     }]
                 }, {
-                    "id":
-                        0,
+                    "id": 0,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            1,
+                        "uid": 1,
+                        "flag": 1,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 20:58:54",
@@ -136,13 +104,10 @@ def fixture_log():
                         }]
                     }]
                 }, {
-                    "id":
-                        0,
+                    "id": 0,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            1,
+                        "uid": 1,
+                        "flag": 1,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 21:00:40",
@@ -153,13 +118,10 @@ def fixture_log():
                         }]
                     }]
                 }, {
-                    "id":
-                        0,
+                    "id": 0,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            1,
+                        "uid": 1,
+                        "flag": 1,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 21:20:03",
@@ -170,13 +132,10 @@ def fixture_log():
                         }]
                     }]
                 }, {
-                    "id":
-                        0,
+                    "id": 0,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            1,
+                        "uid": 1,
+                        "flag": 1,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 21:40:41",
@@ -187,13 +146,10 @@ def fixture_log():
                         }]
                     }]
                 }, {
-                    "id":
-                        2,
+                    "id": 2,
                     "zones": [{
-                        "uid":
-                            1,
-                        "flag":
-                            0,
+                        "uid": 1,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 22:00:02",
@@ -203,10 +159,8 @@ def fixture_log():
                             "realDuration": 1180
                         }]
                     }, {
-                        "uid":
-                            2,
-                        "flag":
-                            0,
+                        "uid": 2,
+                        "flag": 0,
                         "cycles": [{
                             "id": 1,
                             "startTime": "2018-06-02 22:19:42",
@@ -222,14 +176,8 @@ def fixture_log():
     }
 
 
-@pytest.fixture(scope='module')
-def fixture_queue():
-    """Return a /watering/queue response."""
-    return {"queue": []}
-
-
-@pytest.fixture(scope='module')
-def fixture_runs():
+@pytest.fixture()
+def watering_past_json():
     """Return a /watering/past/<DATE>/<DAYS> response."""
     return {
         "pastValues": [{
@@ -292,50 +240,13 @@ def fixture_runs():
     }
 
 
-@pytest.fixture(scope='module')
-def fixture_stop_all():
+@pytest.fixture()
+def watering_queue_json():
+    """Return a /watering/queue response."""
+    return {"queue": []}
+
+
+@pytest.fixture()
+def watering_stopall_json():
     """Return a /watering/stopall response."""
     return {"statusCode": 0, "message": "OK"}
-
-
-@pytest.mark.asyncio
-async def test_endpoints(
-        aresponses, fixture_log, fixture_queue, fixture_runs, fixture_stop_all,
-        event_loop):
-    """Test all endpoints."""
-    today = datetime.date.today()
-    today_str = today.strftime('%Y-%m-%d')
-
-    aresponses.add(
-        '{0}:{1}'.format(TEST_HOST, TEST_PORT),
-        '/api/4/watering/log/details/{0}/{1}'.format(today_str, 2), 'get',
-        aresponses.Response(text=json.dumps(fixture_log), status=200))
-    aresponses.add(
-        '{0}:{1}'.format(TEST_HOST, TEST_PORT), '/api/4/watering/queue', 'get',
-        aresponses.Response(text=json.dumps(fixture_queue), status=200))
-    aresponses.add(
-        '{0}:{1}'.format(TEST_HOST, TEST_PORT),
-        '/api/4/watering/past/{0}/{1}'.format(today_str, 2), 'get',
-        aresponses.Response(text=json.dumps(fixture_runs), status=200))
-    aresponses.add(
-        '{0}:{1}'.format(TEST_HOST, TEST_PORT), '/api/4/watering/stopall',
-        'post',
-        aresponses.Response(text=json.dumps(fixture_stop_all), status=200))
-
-    # pylint: disable=protected-access
-    async with aiohttp.ClientSession(loop=event_loop) as websession:
-        client = Client(TEST_HOST, websession, port=TEST_PORT, ssl=False)
-        client._authenticated = True
-        client._access_token = '12345'
-
-        data = await client.watering.log(today, 2, details=True)
-        assert len(data) == 2
-
-        data = await client.watering.queue()
-        assert not data
-
-        data = await client.watering.runs(today, 2)
-        assert len(data) == 8
-
-        data = await client.watering.stop_all()
-        assert data['message'] == 'OK'
